+++ src/synth-kitchen/contexts/dispatch.ts
@@ -1,8 +1,6 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { IAction } from '../state/actions';
 
 export const DispatchContext = React.createContext<React.Dispatch<IAction>>(
 	() => {}
 );
-
-export const useDispatchContext = () => useContext(DispatchContext);
