+++ src/synth-kitchen/contexts/state.ts
@@ -1,6 +1,4 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { initialState, IState } from '../state/types/state';
 
 export const StateContext = React.createContext<IState>(initialState);
-
-export const useStateContext = () => useContext(StateContext);
