+++ src/synth-kitchen/hooks/use-animation-context.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { AnimationContextInner } from '../contexts/animation';
+
+export const useAnimationContext = () => useContext(AnimationContextInner);
