+++ src/synth-kitchen/hooks/use-dispatch-context.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { DispatchContext } from '../contexts/dispatch';
+
+export const useDispatchContext = () => useContext(DispatchContext);
