+++ src/synth-kitchen/hooks/use-state-context.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { StateContext } from '../contexts/state';
+
+export const useStateContext = () => useContext(StateContext);
